pyproject.toml
[project]
name = "bd-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "google-cloud-storage>=3.10.1",
    "ipykernel>=7.2.0",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.8",
    "pandas>=3.0.1",
    "psycopg2>=2.9.11",
    "pyarrow>=23.0.1",
    "requests>=2.32.5",
    "seaborn>=0.13.2",
]

[dependency-groups]
dev = [
    "missingno>=0.5.2",
    "pgcli>=4.4.0",
]
